Remove os comandos desnecessários
