+++ dsm/extdeps/isis_migration/migration_manager.py
@@ -1354,45 +1354,3 @@ class MigratedDocument:
             remote=remote, local=os.path.basename(zip_file_path))
         self.tracker.info(f"total of zipped files: {len(self.files_to_zip)}")
         return zip_file_path
-
-
-def get_document_pids_to_migrate(from_date=None, to_date=None):
-    """
-    Consulta a base de dados ISIS artigo e retorna os pids atualizados
-    em um intervalo de datas (data de processamento do converter)
-
-    """
-    BASES_ARTIGO_PATH = get_bases_artigo_path()
-    name = date_now_as_folder_name()
-    finished_file_path = create_temp_file(f"{name}_finished.out")
-    output_file_path = create_temp_file(f"{name}_output.csv")
-    from_date = from_date or '0'*8
-    to_date = to_date or '9'*8
-    cmd = (
-        f'''{get_cisis_path()}/ifkeys {BASES_ARTIGO_PATH} '''
-        f'''from=OAITS={from_date} to=OAITS={to_date} > '''
-        f'''{output_file_path};'''
-        f'''echo finished> {finished_file_path}'''
-    )
-    os.system(cmd)
-    while "finished" not in read_file(finished_file_path):
-        pass
-
-    with open(output_file_path, "r") as fp:
-        """
-        output_file content
-
-         1|OAITS=20210917=2352-22912021005005225
-         1|OAITS=20210917=2352-22912021005005226
-         1|OAITS=20210917=2352-22912021005005227
-         1|OAITS=20210917=2352-22912021005005228
-         1|OAITS=20210917=2675-54752021000300400
-         1|OAITS=20210917=2675-54752021000300401
-         1|OAITS=20210917=2675-54752021000300402
-         1|OAITS=20210917=2675-54752021000300700
-
-        """
-        for row in fp:
-            # 1|OAITS=20210917=2675-54752021000300700
-            parts = row.strip().split("=")
-            yield {"updated": parts[1], "pid": "S" + parts[-1]}
